requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# AI/ML Dependencies
openai==1.3.0
google-generativeai==0.3.1
google-genai==0.2.0
numpy==1.24.3
torch>=2.0.0
transformers>=4.30.0

# PDF Processing
PyPDF2==3.0.1

# Utilities
colorama==0.4.6
